AWS_Cloud/src/dags/requirements.txt
apache-airflow-providers-amazon==8.12.0
boto3
